printf: Teach clang about curl_printf formats
Since curl defines printf and friends to curl_mprintf, there are not
warnings for erroneous format specifiers, even though -Wformat is
enabled for clang and gcc.

In order to actually enable warnings from -Wformat for these functions,
add an attribute for those if clang is used as compiler. Unfortunately,
this can not be enabled for gcc, since the c standard curl uses is c89
where %zu and friends are not defined. If gcc sees these attributes, it
will throw warnings about %zu not being included in the c89 standard.